Explicit boxing in main classes
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpclient/trunk@1406891 13f79535-47bb-0310-9956-ffa450edef68

+++ httpclient/src/examples/org/apache/http/examples/client/ClientExecuteSOCKS.java
@@ -104,7 +104,7 @@ public class ClientExecuteSOCKS {
             String proxyHost = (String) params.getParameter("socks.host");
             Integer proxyPort = (Integer) params.getParameter("socks.port");
 
-            InetSocketAddress socksaddr = new InetSocketAddress(proxyHost, proxyPort);
+            InetSocketAddress socksaddr = new InetSocketAddress(proxyHost, proxyPort.intValue());
             Proxy proxy = new Proxy(Proxy.Type.SOCKS, socksaddr);
             return new Socket(proxy);
         }
